third_party/WebKit/LayoutTests/editing/execCommand/print.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable><span>|foo<p>Heres to the crazy ones</p></span></div>',
  'print',
  '<div contenteditable><span>|foo<p>Heres to the crazy ones</p></span></div>'),
  'print');
</script>
